Refactor Ioc to internal Rename Ioc to zeroioc
Expos registermany in zeroapp
