+++ Xam.Zero/Forms/Xam.Zero.Dev/Xam.Zero.Dev/AppShell.xaml.cs
@@ -8,5 +8,7 @@ namespace Xam.Zero.Dev
         {
             this.InitializeComponent();
         }
+        
+        
     }
 }
